% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QA_sim.R
\docType{data}
\name{QA_sim}
\alias{QA_sim}
\title{Name of the file containing phenotypic information}
\format{A data frame 179 genotypes and 3 variables.}
\source{
{MABDE}
}
\usage{
QA_sim
}
\description{
A research program (MABDE) was set up to investigate patterns of
adaptation in barley. In this project a large set of barley genotypes
(~190 genotypes) were evaluated in Europe and in the Mediterranean region.
More details about this population and research can be found
in Comadran et al. (2009).
In this example we simulate the phenotipic data in three environments,
for a set of 179 genotypes. The population was genotyped by DArTs.
}
\keyword{datasets}

